refactor padded_slice to extract each program into their own .hpp/.cpp
