+++ ttnn/cpp/ttnn/operations/experimental/padded_slice/device/padded_slice_rm_program_factory.hpp
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ttnn/tensor/tensor.hpp"
+#include "ttnn/run_operation.hpp"
+
+namespace ttnn::operations::experimental::detail {
+
+tt::tt_metal::operation::ProgramWithCallbacks padded_slice_rm_multi_core(
+    const Tensor& a, Tensor& output, const ttnn::Shape& output_tensor_start, const ttnn::Shape& output_tensor_end);
+
+}  // namespace ttnn::operations::experimental::detail
